Support client reconnects
If the server goes down and comes back up, the client should
automatically try to reconnect. The server already handles stale
clients, so this is the missing half of the puzzle.

The client reconnect is implemented by creating a heartbeat thread.
Every tick, it checks to see the last heartbeat time. If it's been more
than a certain amount of time, it reinitiates the connection.

The logic is spread out over a few different specialized classes. It
isn't as clear as I'd hoped, but that's the price you pay for
"sophisticated" behavior.